Enable sharing `JObject`s across isolates (#1060)
We use a `FinalizableHandle` which is for an entire isolate group as opposed to a `NativeFinalizer` which belongs to a single isolate. This way we can use `@pragma('vm:deeply-immutable')` on `JReference` to share the reference object between multiple isolates while the finalizable handle takes care of the global reference deletion.

The guarantees we provide:
* use-after-free will throw from Dart
* use-on-isolate-b-after-free-on-isolate-a will throw from Dart
* double-free will throw from Dart
* free-on-isolate-b-after-free-on-isolate-a will throw from Dart

In general, if we're not eagerly freeing via `.release()`, things work.